+++ src/tests/App.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import { App } from './App';
+import { App } from '../App';
 
 it('renders "Welcome to Your Fluent UI App"', () => {
   // render(<App />);
